First Project
I was learning the ropes over the course of the project. I started by making the head and worked my way down (which is why the horse head is made much worse than everything else). The naming conventions are bad and my code is messy.  I decided to keep the messy "quirks" of this program in order to view my progress as I make more projects and so I have this state saved as I go forward.